requirements.txt
# MHXY AI Model Factory - Training Environment Dependencies

# 1. 核心深度学习框架 (由ultralytics自动安装)
# torch
# torchvision

# 2. YOLOv8 核心库
ultralytics

# 3. 数据处理与科学计算
numpy
scikit-learn 
tqdm          # 提供优雅的进度条
pyyaml        # 用于解析.yaml配置文件

# 4. ONNX 相关 (用于模型导出和验证)
onnx
onnx-simplifier # 可选，但建议保留用于优化
